refactor: 상영정보 조회 시간 로직 리팩토링 (#34)
* chore: 불필요한 wrapper DTO 삭제

* refactor: 상영정보 조회시 시간 관련 로직 리팩토링
