+++ src/main/resources/application.yml
@@ -15,3 +15,6 @@ spring:
         format_sql: true
         show_sql: true
     database-platform: org.hibernate.dialect.MySQL8Dialect
+
+  jackson:
+    time-zone: Asia/Seoul
